[analyzer] reorganize plotting module structure
