+++ src/canns/analyzer/__init__.py
@@ -1,4 +1,10 @@
-from . import experimental_data as experimental_data
-from . import theta_sweep as theta_sweep
-from . import utils as utils
-from . import visualize as visualize
+"""Analyzer utilities for inspecting CANNS models and simulations."""
+
+from . import experimental_data, plotting, theta_sweep, utils
+
+__all__ = [
+    "experimental_data",
+    "plotting",
+    "theta_sweep",
+    "utils",
+]
