+++ src/canns/analyzer/experimental_data/__init__.py
@@ -1,4 +1,4 @@
-from ..visualize import PlotConfig
+from ..plotting import PlotConfig
 from ._datasets_utils import load_grid_data, load_roi_data, validate_grid_data, validate_roi_data
 from .cann1d import CANN1DPlotConfig, bump_fits, create_1d_bump_animation
 from .cann2d import (
