+++ tests/task/tracking/test_tracking2d.py
@@ -1,7 +1,7 @@
 import brainstate
 import numpy as np
 
-from canns.analyzer.visualize import energy_landscape_2d_animation
+from canns.analyzer.plotting import energy_landscape_2d_animation
 from canns.task.tracking import PopulationCoding2D, TemplateMatching2D, SmoothTracking2D
 from canns.models.basic import CANN2D, CANN2D_SFA
